requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asn1crypto==1.5.1
    # via scramp
boto3==1.34.88
    # via
    #   -r requirements.in
    #   moto
boto3-stubs[essential]==1.34.104
    # via -r requirements.in
botocore==1.34.88
    # via
    #   boto3
    #   moto
    #   s3transfer
botocore-stubs==1.34.94
    # via
    #   -r requirements.in
    #   boto3-stubs
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
coverage[toml]==7.4.4
    # via pytest-cov
cryptography==42.0.5
    # via moto
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via moto
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
moto[s3]==5.0.5
    # via -r requirements.in
mypy-boto3-cloudformation==1.34.84
    # via boto3-stubs
mypy-boto3-dynamodb==1.34.97
    # via boto3-stubs
mypy-boto3-ec2==1.34.101
    # via boto3-stubs
mypy-boto3-lambda==1.34.77
    # via boto3-stubs
mypy-boto3-rds==1.34.93
    # via boto3-stubs
mypy-boto3-s3==1.34.91
    # via boto3-stubs
mypy-boto3-sqs==1.34.101
    # via boto3-stubs
packaging==24.0
    # via pytest
pg8000==1.31.2
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
py-partiql-parser==0.5.4
    # via moto
pycparser==2.22
    # via cffi
pytest==8.1.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-testdox
pytest-cov==5.0.0
    # via -r requirements.in
pytest-testdox==3.1.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   moto
    #   pg8000
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   moto
    #   responses
requests==2.31.0
    # via
    #   -r requirements.in
    #   moto
    #   responses
responses==0.25.0
    # via moto
s3transfer==0.10.1
    # via boto3
scramp==1.4.5
    # via pg8000
six==1.16.0
    # via python-dateutil
types-awscrt==0.20.9
    # via botocore-stubs
types-s3transfer==0.10.1
    # via boto3-stubs
typing-extensions==4.11.0
    # via
    #   boto3-stubs
    #   mypy-boto3-cloudformation
    #   mypy-boto3-dynamodb
    #   mypy-boto3-ec2
    #   mypy-boto3-lambda
    #   mypy-boto3-rds
    #   mypy-boto3-s3
    #   mypy-boto3-sqs
urllib3==2.2.1
    # via
    #   botocore
    #   requests
    #   responses
werkzeug==3.0.2
    # via moto
xmltodict==0.13.0
    # via moto
